refactor: move frontend options to common_options
