+++ src/cmd/src/frontend.rs
@@ -16,11 +16,10 @@ use std::sync::Arc;
 
 use clap::Parser;
 use common_base::Plugins;
+use common_options::frontend::{FrontendOptions, InfluxdbOptions, PrometheusOptions};
+use common_options::meta::MetaClientOptions;
 use common_telemetry::logging;
-use frontend::frontend::FrontendOptions;
 use frontend::instance::{FrontendInstance, Instance as FeInstance};
-use frontend::service_config::{InfluxdbOptions, PrometheusOptions};
-use meta_client::MetaClientOptions;
 use servers::auth::UserProviderRef;
 use servers::tls::{TlsMode, TlsOption};
 use servers::{auth, Mode};
@@ -237,8 +236,8 @@ mod tests {
     use std::time::Duration;
 
     use common_base::readable_size::ReadableSize;
+    use common_options::frontend::GrpcOptions;
     use common_test_util::temp_dir::create_named_temp_file;
-    use frontend::service_config::GrpcOptions;
     use servers::auth::{Identity, Password, UserProviderRef};
 
     use super::*;
