+++ src/frontend/Cargo.toml
@@ -26,6 +26,7 @@ common-grpc = { path = "../common/grpc" }
 common-grpc-expr = { path = "../common/grpc-expr" }
 common-query = { path = "../common/query" }
 common-meta = { path = "../common/meta" }
+common-options = { path = "../common/options" }
 common-recordbatch = { path = "../common/recordbatch" }
 common-runtime = { path = "../common/runtime" }
 common-telemetry = { path = "../common/telemetry" }
