+++ src/frontend/src/lib.rs
@@ -18,12 +18,10 @@
 pub mod catalog;
 pub mod error;
 pub mod expr_factory;
-pub mod frontend;
 pub mod heartbeat;
 pub mod instance;
 pub(crate) mod metrics;
 mod script;
 mod server;
-pub mod service_config;
 pub mod statement;
 pub mod table;
