+++ tests-integration/src/cluster.rs
@@ -23,13 +23,13 @@ use common_base::Plugins;
 use common_grpc::channel_manager::ChannelManager;
 use common_meta::peer::Peer;
 use common_meta::DatanodeId;
+use common_options::frontend::FrontendOptions;
 use common_options::meta::MetaSrvOptions;
 use common_runtime::Builder as RuntimeBuilder;
 use common_test_util::temp_dir::create_temp_dir;
 use datanode::datanode::{DatanodeOptions, ObjectStoreConfig, ProcedureConfig};
 use datanode::heartbeat::HeartbeatTask;
 use datanode::instance::Instance as DatanodeInstance;
-use frontend::frontend::FrontendOptions;
 use frontend::instance::{FrontendInstance, Instance as FeInstance};
 use meta_client::client::MetaClientBuilder;
 use meta_srv::cluster::MetaPeerClientRef;
